+++ src/main/java/com/devvictor/ecommerce_api/application/dtos/output/CartOutputDTO.java
@@ -6,6 +6,6 @@ public record CartOutputDTO(
         String id,
         List<CartProductOutputDTO> products,
         int productsQuantity,
-        double totalPrice
+        long totalPrice
 ) {
 }
