+++ src/main/java/com/devvictor/ecommerce_api/application/dtos/output/CartProductOutputDTO.java
@@ -2,7 +2,7 @@ package com.devvictor.ecommerce_api.application.dtos.output;
 
 public record CartProductOutputDTO(
         String id,
-        double price,
+        long price,
         String description,
         String photoUrl,
         int inCartQuantity
