+++ src/main/java/com/devvictor/ecommerce_api/application/dtos/output/ProductOutputDTO.java
@@ -2,7 +2,7 @@ package com.devvictor.ecommerce_api.application.dtos.output;
 
 public record ProductOutputDTO(
         String id,
-        double price,
+        long price,
         String name,
         String description,
         String photoUrl,
